+++ apps/backend/src/migrator/m20230419_000003_create_seen.rs
@@ -1,6 +1,6 @@
 use sea_orm_migration::prelude::*;
 
-use super::{m20230417_000002_create_user::User, Metadata};
+use crate::migrator::{m20230417_000002_create_user::User, Metadata};
 
 pub struct Migration;
