+++ apps/backend/src/migrator/m20230502_000004_create_genre.rs
@@ -1,6 +1,6 @@
 use sea_orm_migration::prelude::*;
 
-use super::Metadata;
+use crate::migrator::Metadata;
 
 pub struct Migration;
