+++ apps/backend/src/migrator/m20230507_000007_create_collection.rs
@@ -1,6 +1,6 @@
 use sea_orm_migration::prelude::*;
 
-use super::{m20230417_000002_create_user::User, Metadata};
+use crate::migrator::{m20230417_000002_create_user::User, Metadata};
 
 pub struct Migration;
