+++ apps/backend/src/migrator/m20230621_000012_add_metadata_unique_index.rs
@@ -1,6 +1,6 @@
 use sea_orm_migration::prelude::*;
 
-use super::{
+use crate::migrator::{
     m20230410_000001_create_metadata::{MetadataImage, UNIQUE_INDEX},
     Metadata,
 };
